requirements.txt
asgiref==3.5.2
celery==5.2.7
Django==4.1
djangorestframework==3.13.1
django-debug-toolbar==3.6.0
django-colorfield==0.7.2
django-admin-interface==0.20.0
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2